requirements.txt
google-cloud-storage==2.10.0
google-cloud-bigquery==3.13.0
google-cloud-secret-manager==2.16.0
gspread==6.1.0
gspread-dataframe==3.3.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
pandas==2.1.3
numpy==1.24.3
python-dotenv==1.0.0
openai==1.3.0
openpyxl==3.1.5
anthropic==0.7.0
functions-framework==3.5.0
streamlit==1.32.0
email-validator==2.1.0
google-generativeai==0.3.0
